+++ CHANGELOG.md
@@ -46,6 +46,7 @@ All notable changes to this project will be documented in this file.
 - Model.GetElements() method
 - Remove SL4 support (#115)
 - Remove NET35 support (#115)
+- PlotElement.Format method. Use StringHelper.Format instead.
 
 ### Fixed
 - Tracker position is wrong when PlotView is offset from origin (#455)